+++ src/libs/ascent/runtimes/ascent_vtkh_data_adapter.hpp
@@ -167,6 +167,11 @@ private:
                                      const std::string &topo_name,
                                      bool zero_copy);
 
+    static void VTKmBlueprintShapeMap(conduit::Node &output);
+
+    static bool CheckShapeMapVsVTKmShapeIds(const conduit::Node &shape_map);
+
+
 };
 
 //-----------------------------------------------------------------------------